moar img
rem was corrreeect but i used a rando number from somewhere else in the sheet and that was not de wey, switched to 20. randomly. on a whim. Looks ok in my desktop mobile view but that shit has been off every time lmfao.

asked chatGPT to fix the scroll glitch w my mobile header background, realized the URL bar was eating the VH so it did fancy math for me. idk whatll happen. hehe cya